requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.0.0
django-storages==1.12.3
boto3==1.26.0
stripe==8.5.0
drf-spectacular==0.25.0
mysqlclient